Embed staging into driip settlement challenge and defer timeout
Add stage amounts into start challenge functions. Challenge can not be
started if stage amount is greater than balance. By subtracting stage
amount from (current) balance the target balance that is the value
to which driip settlement challenge candidates are compared in challenge
functions.

In the event that a driip settlement is challenged by an order the
timer of the ongoing challenge is reset to allow for ample time to
unchallenge the challenge order.