Update the calculation of NSC proposal nonce
Calculate the new NSC proposal nonce as the max of the DSC nonce and
(former) NSC nonce for the concerned wallet/currency pair.

In stopChallenge() and stopChallenteByProxy() terminate proposal rather
than remove it.

Remove the reference to NullSettlementState.

Add partial require on non-existence of NSC proposal at start of
challenge.

Add functions hasProposal() and hasProposalTerminated().