Support claimable amount getter and claim by accrual period index range
Add function claimableAmount() that returns the amount that can be
claimed by a wallet-currency pair for a range of accrual period indices.

Add accrual period index boundary params to function claimAndStage() to
claim for a range of accrual periods.

Overload function claimAndTransferToBeneficiary() with accrual period
index boundary params to claim for a range of accrual periods.